+++ RioSharp/RioSocketPool.cs
@@ -44,11 +44,11 @@ namespace RioSharp
                 Imports.ThrowLastError();
 
 
-            _sendBufferId = RioStatic.RegisterBuffer(SendBufferPool.BufferPointer, SendBufferPool.TotalLength);
+            _sendBufferId = RioStatic.RegisterBuffer(SendBufferPool.BufferPointer, (uint)SendBufferPool.TotalLength);
             Imports.ThrowLastWSAError();
             SendBufferPool.SetBufferId(_sendBufferId);
 
-            _reciveBufferId = RioStatic.RegisterBuffer(ReciveBufferPool.BufferPointer, ReciveBufferPool.TotalLength);
+            _reciveBufferId = RioStatic.RegisterBuffer(ReciveBufferPool.BufferPointer, (uint)ReciveBufferPool.TotalLength);
             Imports.ThrowLastWSAError();
             ReciveBufferPool.SetBufferId(_reciveBufferId);
 
@@ -63,7 +63,7 @@ namespace RioSharp
                 }
             };
 
-            if ((SendCompletionQueue = RioStatic.CreateCompletionQueue((uint)MaxOutsandingCompletions, sendCompletionMethod)) == IntPtr.Zero)
+            if ((SendCompletionQueue = RioStatic.CreateCompletionQueue(MaxOutsandingCompletions, sendCompletionMethod)) == IntPtr.Zero)
                 Imports.ThrowLastWSAError();
 
             var receiveCompletionMethod = new RIO_NOTIFICATION_COMPLETION()
@@ -77,7 +77,7 @@ namespace RioSharp
                 }
             };
 
-            if ((ReceiveCompletionQueue = RioStatic.CreateCompletionQueue((uint)MaxOutsandingCompletions, receiveCompletionMethod)) == IntPtr.Zero)
+            if ((ReceiveCompletionQueue = RioStatic.CreateCompletionQueue(MaxOutsandingCompletions, receiveCompletionMethod)) == IntPtr.Zero)
                 Imports.ThrowLastWSAError();
 
 
@@ -97,7 +97,7 @@ namespace RioSharp
                 Buffer.MemoryCopy(p, currentSegment.rawPointer, SendBufferPool.SegmentLength, buffer.Length);
             }
             
-            currentSegment.segmentPointer->Length = (uint)buffer.Length; 
+            currentSegment.segmentPointer->Length = buffer.Length; 
             currentSegment.AutoFree = false;
             return currentSegment;
         }
@@ -130,7 +130,7 @@ namespace RioSharp
                             buf = ReciveBufferPool.allSegments[result.RequestCorrelation];
                             if (connections.TryGetValue(result.ConnectionCorrelation, out connection))
                             {
-                                buf.segmentPointer->Length = result.BytesTransferred; //use currentLength instead?
+                                buf.segmentPointer->Length = (int)result.BytesTransferred; //use currentLength instead?
                                 connection.incommingSegments.Set(buf);
                             }
                             else
